requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
gallery_dl==1.31.4
greenlet==3.3.1
h11==0.16.0
idna==3.11
instaloader==4.15
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.0
playwright==1.57.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
starlette==0.50.0
tiktok-dlpy==2.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
yt-dlp==2025.12.8
